Add usbr-git-library to perform Git Config setup
Recent versions of Git for Windows have begun supporting OAuth login
for self hosted Gitlab instaces. This breaks our existing login
workflow that involved creating a Personal Access Token and feeding it
to Git.

When properly configured, this new workflow is more convenient (as long
as the user is already signed in to Gitlab, it's a two-click operation.
No more access tokens for Git).

The downside is that each Gitlab much be configured for access, and
every Git client machine must be configured for each Gitlab instance.

The new Git library supports utilizing the Git CLI tools (must be on the
PATH) to read the Git config, as well as to update it. The values needed
for making Gitlab work may be specified in an XML file. That XML file
may be bundled into a Jar, which makes it possible to bundle the
configuration within WTMP in a convenient manner. Multiple Gitlab
configurations may be specified within the XML file, permitting multiple
Gitlab instances to be auto-configred. Additionally, the Git tooling
incorporates some wtmp-specific properties to tell WTMP to ignore the
properties, in the case that a user does not want their system
automatically reconfigured by WTMP.